+++ video-call/src/main/java/com/nabla/sdk/videocall/ReporterExtensions.kt
@@ -1,0 +1,6 @@
+package com.nabla.sdk.videocall
+
+import com.nabla.sdk.core.domain.boundary.ErrorReporter
+
+internal val ErrorReporter.Companion.VIDEO_CALL_DOMAIN: String
+    get() = "video-call"