+++ video-call/src/main/java/com/nabla/sdk/videocall/injection/VideoCallContainer.kt
@@ -1,9 +1,11 @@
 package com.nabla.sdk.videocall.injection
 
 import android.content.Context
+import com.nabla.sdk.core.domain.boundary.ErrorReporter
 import com.nabla.sdk.core.domain.boundary.Logger
 import com.nabla.sdk.videocall.data.CameraServiceImpl
 import com.nabla.sdk.videocall.data.VideoCallRepository
+import com.nabla.sdk.videocall.data.VideoCallRepositoryReporterHelper
 import com.nabla.sdk.videocall.domain.CameraService
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -14,6 +16,7 @@ internal class VideoCallContainer(
     baseOkHttpClient: OkHttpClient,
     val logger: Logger,
     val context: Context,
+    val errorReporter: ErrorReporter,
 ) {
 
     private val okHttpClient = baseOkHttpClient.newBuilder()
@@ -23,7 +26,7 @@ internal class VideoCallContainer(
         applicationContext = context,
         okHttpClient = okHttpClient,
         repoScope = CoroutineScope(SupervisorJob() + Dispatchers.IO),
-        logger = logger,
+        videoCallRepositoryReporterHelper = VideoCallRepositoryReporterHelper(errorReporter)
     )
 
     val cameraService: CameraService = CameraServiceImpl(logger, context.applicationContext)